Add ship turn scenario and fix update_hp bug

+++ app/navy/services/ship_service.py
@@ -120,7 +120,8 @@ class ShipService:
     def update_hp(self, ship, damage):
         if ship.hp - damage <= utils.ZERO:
             self.delete(ship)
-            self.delete_from_board(ship)
+            if self.is_alive_on_board(ship):
+                self.delete_from_board(ship)
         else:
             ship.hp -= damage
             ship_dao.add_or_update(ship)
@@ -137,6 +138,12 @@ class ShipService:
     # -- Private Methods -- #
     def is_alive(self, ship_id):
         return ship_dao.get_by_id(ship_id)
+
+    # Como para evitar buscar posicion por posicion en el mapa
+    def is_alive_on_board(self, ship):
+        from app.navy.services.navy_game_service import navy_game_service
+
+        return ship in navy_game_service.games[ship.navy_game_id]["ships"]
 
     def act_accordingly_to_ship(self, ship, other_ship):
         old_hp = ship.hp
